# legacy/python-architecture-0002/agent/inspector/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Memory Observatory API",
    "description": "Read-only LongMemEval temporal graph artifact inspector.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/runs": {
      "get": {
        "summary": "List Runs",
        "operationId": "list_runs_api_runs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunSummary"
                  },
                  "type": "array",
                  "title": "Response List Runs Api Runs Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/cases": {
      "get": {
        "summary": "List Cases",
        "operationId": "list_cases_api_runs__run_id__cases_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CaseSummary"
                  },
                  "title": "Response List Cases Api Runs  Run Id  Cases Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/cases/{case_id}": {
      "get": {
        "summary": "Get Case",
        "operationId": "get_case_api_runs__run_id__cases__case_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "batch",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Batch"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/cases/{case_id}/export.svg": {
      "get": {
        "summary": "Export Case Svg",
        "operationId": "export_case_svg_api_runs__run_id__cases__case_id__export_svg_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "batch",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Batch"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/cases/{case_id}/events": {
      "get": {
        "summary": "Stream Case Events",
        "operationId": "stream_case_events_api_runs__run_id__cases__case_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CaseSnapshot": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "graph": {
            "$ref": "#/components/schemas/JsonObject"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/JsonObject"
            },
            "type": "array",
            "title": "Events"
          },
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/JsonObject"
            },
            "type": "array",
            "title": "Sessions"
          },
          "model_calls": {
            "items": {
              "$ref": "#/components/schemas/JsonObject"
            },
            "type": "array",
            "title": "Model Calls"
          },
          "answer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JsonObject"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "case_id",
          "graph",
          "events",
          "sessions",
          "model_calls",
          "status"
        ],
        "title": "CaseSnapshot"
      },
      "CaseSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "event_count": {
            "type": "integer",
            "title": "Event Count"
          },
          "batch_count": {
            "type": "integer",
            "title": "Batch Count"
          },
          "has_final_graph": {
            "type": "boolean",
            "title": "Has Final Graph"
          },
          "has_answer": {
            "type": "boolean",
            "title": "Has Answer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "event_count",
          "batch_count",
          "has_final_graph",
          "has_answer"
        ],
        "title": "CaseSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JsonObject": {
        "additionalProperties": {
          "$ref": "#/components/schemas/JsonValue"
        },
        "type": "object"
      },
      "JsonScalar": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "type": "null"
          }
        ]
      },
      "JsonValue": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/JsonScalar"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "array"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object"
          }
        ]
      },
      "RunSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "relative_path": {
            "type": "string",
            "title": "Relative Path"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "architecture": {
            "type": "string",
            "title": "Architecture"
          },
          "completed_count": {
            "type": "integer",
            "title": "Completed Count",
            "default": 0
          },
          "selected_count": {
            "type": "integer",
            "title": "Selected Count",
            "default": 0
          },
          "has_graph_artifacts": {
            "type": "boolean",
            "title": "Has Graph Artifacts",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "relative_path",
          "status",
          "architecture"
        ],
        "title": "RunSummary"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}